Rebase the code

+++ Finished/Aula02_09/HomeWork2.java
@@ -1,6 +1,10 @@
 public class HomeWork2 {
+
+    // use double values, implict that the max number is 170
+    // using float values, implict that the max number is 80
+    // using long values, implict that the max number is 30
     
-    static long factorial(int n) {
+    static double factorial(int n) {
         if (n == 0)
             return 1;
         if (n < 0)
@@ -15,7 +19,7 @@ public class HomeWork2 {
             number = Integer.parseInt(args[0]);
         }
 
-        long factorial;
+        double factorial;
         factorial = factorial(number);
         System.out.println("The factorial of " + number + " is " + factorial);
