+++ Finished/Aula02_09/HomeWork3.java
@@ -1,6 +1,6 @@
 public class HomeWork3 {
 
-    static long factorial(int n) {
+    static double factorial(int n) {
         if (n == 0)
             return 1;
         if (n < 0)
@@ -13,9 +13,9 @@ public class HomeWork3 {
         int number = Integer.parseInt(args[0]);
         int number2 = Integer.parseInt(args[1]);
 
-        long factorial = factorial(number);
-        long factorial2 = factorial(number2);
-        long result = factorial + factorial2;
+        double factorial = factorial(number);
+        double factorial2 = factorial(number2);
+        double result = factorial + factorial2;
         System.out.println("The factorial between the numbers " + number + " and " + number2
                 + " is " + result);
     }
